Restructure intent handling
