+++ applications/ffd/src/intent_handler/intent_handler.h
@@ -1,0 +1,9 @@
+// Copyright (c) 2022 XMOS LIMITED. This Software is subject to the terms of the
+// XMOS Public License: Version 1
+
+#ifndef INTENT_HANDLER_H_
+#define INTENT_HANDLER_H_
+
+int32_t intent_handler_create(uint32_t priority, void *args);
+
+#endif /* INTENT_HANDLER_H_ */
