+++ applications/ffd/src/main.c
@@ -28,6 +28,7 @@
 #include "rtos_swmem.h"
 #include "xcore_device_memory.h"
 #include "ssd1306_rtos_support.h"
+#include "intent_handler/intent_handler.h"
 
 extern void startup_task(void *arg);
 extern void tile_common_init(chanend_t c);
@@ -99,7 +100,9 @@ void startup_task(void *arg)
 #if appconfSSD1306_DISPLAY_ENABLED
     ssd1306_display_create(appconfSSD1306_TASK_PRIORITY);
 #endif
-    inference_engine_create(appconfINFERENCE_MODEL_RUNNER_TASK_PRIORITY, NULL);
+    QueueHandle_t q_intent = xQueueCreate(appconfINTENT_QUEUE_LEN, sizeof(int32_t));
+    intent_handler_create(appconfINFERENCE_MODEL_RUNNER_TASK_PRIORITY, q_intent);
+    inference_engine_create(appconfINFERENCE_MODEL_RUNNER_TASK_PRIORITY, q_intent);
 #endif
 
 #if ON_TILE(AUDIO_PIPELINE_TILE_NO)
